Extract AI co-author detection from calculator.py to _ai_detection.py
calculator.py drops from 190 to 150 lines. AI detection is a
self-contained concern (regex patterns + two functions) that
doesn't share state with the time/size metric calculations.
