+++ tests/unit/metrics/test_metrics.py
@@ -810,37 +810,37 @@ class TestIsAiCoauthored:
     """Test cases for is_ai_coauthored function."""
 
     def test_should_return_false_for_none_body(self):
-        from git_dev_metrics.metrics.calculator import is_ai_coauthored
+        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
 
         result = is_ai_coauthored(any_pr(body=None))
         assert result is False
 
     def test_should_return_false_for_empty_body(self):
-        from git_dev_metrics.metrics.calculator import is_ai_coauthored
+        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
 
         result = is_ai_coauthored(any_pr(body=""))
         assert result is False
 
     def test_should_return_false_for_no_trailer(self):
-        from git_dev_metrics.metrics.calculator import is_ai_coauthored
+        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
 
         result = is_ai_coauthored(any_pr(body="This is a regular PR description"))
         assert result is False
 
     def test_should_return_true_for_coauthored_by_in_body(self):
-        from git_dev_metrics.metrics.calculator import is_ai_coauthored
+        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
 
         result = is_ai_coauthored(any_pr(body="Co-Authored-By: GitHub <noreply@github.com>"))
         assert result is True
 
     def test_should_be_case_insensitive(self):
-        from git_dev_metrics.metrics.calculator import is_ai_coauthored
+        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
 
         result = is_ai_coauthored(any_pr(body="co-authored-by: someone@example.com"))
         assert result is True
 
     def test_should_return_true_when_trailer_only_in_commit_message(self):
-        from git_dev_metrics.metrics.calculator import is_ai_coauthored
+        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
 
         pr = any_pr(
             body=None,
@@ -850,7 +850,7 @@ class TestIsAiCoauthored:
         assert result is True
 
     def test_should_return_true_when_trailer_in_one_of_many_commits(self):
-        from git_dev_metrics.metrics.calculator import is_ai_coauthored
+        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
 
         pr = any_pr(
             body="Regular PR",
@@ -864,7 +864,7 @@ class TestIsAiCoauthored:
         assert result is True
 
     def test_should_return_false_when_no_trailer_in_body_or_commits(self):
-        from git_dev_metrics.metrics.calculator import is_ai_coauthored
+        from git_dev_metrics.metrics._ai_detection import is_ai_coauthored
 
         pr = any_pr(
             body="Regular PR",
@@ -878,13 +878,13 @@ class TestCalculateAiPercentage:
     """Test cases for calculate_ai_percentage function."""
 
     def test_should_return_zero_for_empty_list(self):
-        from git_dev_metrics.metrics.calculator import calculate_ai_percentage
+        from git_dev_metrics.metrics._ai_detection import calculate_ai_percentage
 
         result = calculate_ai_percentage([])
         assert result == 0.0
 
     def test_should_return_zero_for_no_ai_prs(self):
-        from git_dev_metrics.metrics.calculator import calculate_ai_percentage
+        from git_dev_metrics.metrics._ai_detection import calculate_ai_percentage
 
         prs = [
             any_pr(body=None),
@@ -895,7 +895,7 @@ class TestCalculateAiPercentage:
         assert result == 0.0
 
     def test_should_calculate_percentage_correctly(self):
-        from git_dev_metrics.metrics.calculator import calculate_ai_percentage
+        from git_dev_metrics.metrics._ai_detection import calculate_ai_percentage
 
         prs = [
             any_pr(body="Co-Authored-By: someone@example.com"),
@@ -907,7 +907,7 @@ class TestCalculateAiPercentage:
         assert result == 50.0
 
     def test_should_handle_all_ai_prs(self):
-        from git_dev_metrics.metrics.calculator import calculate_ai_percentage
+        from git_dev_metrics.metrics._ai_detection import calculate_ai_percentage
 
         prs = [
             any_pr(body="Co-Authored-By: someone@example.com"),
@@ -917,7 +917,7 @@ class TestCalculateAiPercentage:
         assert result == 100.0
 
     def test_should_round_to_one_decimal(self):
-        from git_dev_metrics.metrics.calculator import calculate_ai_percentage
+        from git_dev_metrics.metrics._ai_detection import calculate_ai_percentage
 
         prs = [
             any_pr(body="Co-Authored-By: someone@example.com"),
@@ -928,7 +928,7 @@ class TestCalculateAiPercentage:
         assert result == 33.3
 
     def test_should_count_pr_with_trailer_only_in_commit_message(self):
-        from git_dev_metrics.metrics.calculator import calculate_ai_percentage
+        from git_dev_metrics.metrics._ai_detection import calculate_ai_percentage
 
         prs = [
             any_pr(body="Regular PR", commit_messages=["Co-Authored-By: Claude"]),
